requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage==1.16.1
numpy==1.17.4
gensim==3.8.1
nltk==3.4.5
#torch download
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
spacy>=2.2.0,<3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm